Added .gitignore, Structured files, edited README.md
Now have gitignore. Use this to notate files you would not like synced to github and only want for your personal machine. This will be helpful when we start putting code on here but I wanted a notes section to keep with the repo but not to upload as they're just casual notes.